Update 08.04.18
1. Fixed bugs with instantiate bullet and sync with another clients.
2. Recreated Register and Unregister systems.
3. Created take damage system

+++ Survival2D/Assets/PlayerWeapon.cs
@@ -5,7 +5,7 @@ public class PlayerWeapon {
 
     public string name = "Glock";
 
-    public float damage = 23f;
+    public int damage = 23;
     public float range = 70f;
     public float speed = 10f;
